Personnel management improvements.

+++ src/trpc/routers/team-memberships.ts
@@ -3,15 +3,17 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
 */
 
+import { pick } from 'remeda'
 import { match } from 'ts-pattern'
 import { z } from 'zod'
 
-import { clerkClient } from '@clerk/nextjs/server'
+import { clerkClient, createClerkClient } from '@clerk/nextjs/server'
 import { TeamMembershipRole } from '@prisma/client'
 import { TRPCError } from '@trpc/server'
 
 import { teamMembershipFormSchema } from '@/lib/forms/team-membership'
-import { zodDeleteType, zodRecordStatus, zodShortId } from '@/lib/validation'
+import { nanoId16 } from '@/lib/id'
+import { zodDeleteType, zodRecordStatus, zodNanoId8 } from '@/lib/validation'
 
 import { AuthenticatedContext, createTRPCRouter, systemAdminProcedure } from '../init'
 import { TeamMembershipWithPerson, TeamMembershipWithPersonAndTeam, TeamMembershipWithTeam } from '../types'
@@ -26,14 +28,15 @@ export const teamMembershipsRouter = createTRPCRouter({
 
             // Check if the team member already exists
             const existing = await ctx.prisma.teamMembership.findFirst({ where: { teamId, personId }, include: { person: true, team: true } })
-            if(existing) throw new TRPCError({ code: 'CONFLICT', message: `Person(${existing.person.name}) is already a member of Team(${existing.team.name}).` })
+            if(existing?.status == 'Active' || existing?.status == 'Inactive') 
+                throw new TRPCError({ code: 'CONFLICT', message: `'${existing.person.name}' is already a member of '${existing.team.name}'.` })
 
-            return addTeamMember(ctx, input)
+            return addTeamMember(ctx, { teamId, personId, role: input.role, restoreId: existing?.id })
         }),
 
     byPerson: systemAdminProcedure
         .input(z.object({
-            personId: zodShortId,
+            personId: zodNanoId8,
             status: zodRecordStatus
         }))
         .query(async ({ ctx, input }): Promise<TeamMembershipWithTeam[]> => {
@@ -49,7 +52,7 @@ export const teamMembershipsRouter = createTRPCRouter({
 
     byTeam: systemAdminProcedure
         .input(z.object({
-            teamId: zodShortId,
+            teamId: zodNanoId8,
             status: zodRecordStatus
         }))
         .query(async ({ ctx, input }): Promise<TeamMembershipWithPerson[]> => {
@@ -65,8 +68,8 @@ export const teamMembershipsRouter = createTRPCRouter({
 
     delete: systemAdminProcedure
         .input(z.object({
-            teamId: zodShortId,
-            personId: zodShortId,
+            teamId: zodNanoId8,
+            personId: zodNanoId8,
             deleteType: zodDeleteType
         }))
         .mutation(async ({ ctx, input }) => {
@@ -77,6 +80,8 @@ export const teamMembershipsRouter = createTRPCRouter({
             if(!existing) throw new TRPCError({ code: 'NOT_FOUND' })
 
             await removeTeamMember(ctx, { membership: existing, deleteType: input.deleteType })
+
+            return pick(existing, ['id', 'teamId', 'personId', 'role'])
         }),
 
     update: systemAdminProcedure
@@ -101,20 +106,30 @@ export const teamMembershipsRouter = createTRPCRouter({
  * @param teamId The team ID.
  * @param personId The person ID.
  * @param role The role to assign to the person.
+ * @param restoreId The ID of the membership to restore, if applicable.
  * @returns The created team membership.
  */
-export async function addTeamMember(ctx: AuthenticatedContext, { teamId, personId, role }: { teamId: string, personId: string, role: TeamMembershipRole }): Promise<TeamMembershipWithPersonAndTeam> {
+export async function addTeamMember(ctx: AuthenticatedContext, { teamId, personId, role, restoreId }: { teamId: string, personId: string, role: TeamMembershipRole, restoreId: string | undefined }): Promise<TeamMembershipWithPersonAndTeam> {
+    
     const [createdMembership] = await ctx.prisma.$transaction([
-        ctx.prisma.teamMembership.create({
-            data: {
-                team: { connect: { id: teamId }},
-                person: { connect: { id: personId }},
-                role: role,
-            },
-            include: { person: true, team: true, d4hInfo: true }
-        }),
+        restoreId
+            ? ctx.prisma.teamMembership.update({
+                where: { id: restoreId },
+                data: { status: 'Active', role },
+                include: { person: true, team: true, d4hInfo: true }
+            })
+            : ctx.prisma.teamMembership.create({
+                data: {
+                    id: nanoId16(),
+                    team: { connect: { id: teamId }},
+                    person: { connect: { id: personId }},
+                    role: role,
+                },
+                include: { person: true, team: true, d4hInfo: true }
+            }),
         ctx.prisma.teamChangeLog.create({
             data: {
+                id: nanoId16(),
                 teamId,
                 actorId: ctx.personId,
                 event: 'AddMember',
@@ -125,7 +140,7 @@ export async function addTeamMember(ctx: AuthenticatedContext, { teamId, personI
 
     if(createdMembership.person.clerkUserId && role != 'None') {
         // Add the person to the clerk organization
-        const clerk = await clerkClient()
+        const clerk = createClerkClient({ secretKey: process.env.CLERK_SECRET_KEY })
         await clerk.organizations.createOrganizationMembership({
             organizationId: createdMembership.team.clerkOrgId,
             userId: createdMembership.person.clerkUserId,
@@ -144,12 +159,16 @@ export async function addTeamMember(ctx: AuthenticatedContext, { teamId, personI
  */
 export async function removeTeamMember(ctx: AuthenticatedContext, { membership, deleteType }: { membership: TeamMembershipWithPersonAndTeam, deleteType: 'Soft' | 'Hard' }): Promise<void> {
 
+    const person = await ctx.prisma.person.findUnique({ where: { id: membership.personId } })
+    if(!person) throw new TRPCError({ code: 'NOT_FOUND', message: `Person(${membership.personId}) not found.` })
+
     await match(deleteType)
         .with('Hard', () => 
             ctx.prisma.$transaction([
                 ctx.prisma.teamMembership.delete({ where: { id: membership.id }}),
                 ctx.prisma.teamChangeLog.create({
                     data: {
+                        id: nanoId16(),
                         teamId: membership.teamId,
                         actorId: ctx.personId,
                         event: 'RemoveMember',
@@ -166,6 +185,7 @@ export async function removeTeamMember(ctx: AuthenticatedContext, { membership, 
                 }),
                 ctx.prisma.teamChangeLog.create({
                     data: {
+                        id: nanoId16(),
                         teamId: membership.teamId,
                         actorId: ctx.personId,
                         event: 'RemoveMember',
@@ -176,12 +196,12 @@ export async function removeTeamMember(ctx: AuthenticatedContext, { membership, 
         )
         .exhaustive()
 
-    if(membership.role != 'None' && membership.person.clerkUserId) {
+    if(membership.role != 'None' && person.clerkUserId) {
         // Remove the person from the clerk organization
-        const clerk = await clerkClient()
+        const clerk = createClerkClient({ secretKey: process.env.CLERK_SECRET_KEY })
         await clerk.organizations.deleteOrganizationMembership({
             organizationId: membership.team.clerkOrgId,
-            userId: membership.person.clerkUserId
+            userId: person.clerkUserId
         })
     }
 }
@@ -195,7 +215,7 @@ export async function removeTeamMember(ctx: AuthenticatedContext, { membership, 
  */
 export async function updateTeamMember(ctx: AuthenticatedContext, { membership, newRole }: { membership: TeamMembershipWithPersonAndTeam, newRole: TeamMembershipRole }): Promise<TeamMembershipWithPersonAndTeam> {
 
-    const [updatedPerson] = await ctx.prisma.$transaction([
+    const [updatedMembership] = await ctx.prisma.$transaction([
         ctx.prisma.teamMembership.update({
             where: { id: membership.id },
             data: { role: newRole },
@@ -203,6 +223,7 @@ export async function updateTeamMember(ctx: AuthenticatedContext, { membership, 
         }),
         ctx.prisma.teamChangeLog.create({
             data: {
+                id: nanoId16(),
                 teamId: membership.teamId,
                 actorId: ctx.personId,
                 event: 'UpdateMember',
@@ -211,32 +232,35 @@ export async function updateTeamMember(ctx: AuthenticatedContext, { membership, 
         })
     ])
 
-    if(membership.person.clerkUserId) {
+    const person = await ctx.prisma.person.findUnique({ where: { id: membership.personId } })
+    if(!person) throw new TRPCError({ code: 'NOT_FOUND', message: `Person(${membership.personId}) not found.` })
+
+    if(person.clerkUserId) {
         // Update the person's role in the clerk organization
-        const clerk = await clerkClient()
+        const clerk = createClerkClient({ secretKey: process.env.CLERK_SECRET_KEY })
 
         if(newRole == 'None') {
             // Remove from organization
             await clerk.organizations.deleteOrganizationMembership({
                 organizationId: membership.team.clerkOrgId,
-                userId: membership.person.clerkUserId
+                userId: person.clerkUserId
             })
         } else if(membership.role == 'None') {
             // Add to organization
             await clerk.organizations.createOrganizationMembership({
                 organizationId: membership.team.clerkOrgId,
-                userId: membership.person.clerkUserId,
+                userId: person.clerkUserId,
                 role: newRole == 'Admin' ? 'org:admin' : 'org:member'
             })
         } else {
             // Change role
             await clerk.organizations.updateOrganizationMembership({
                 organizationId: membership.team.clerkOrgId,
-                userId: membership.person.clerkUserId,
+                userId: person.clerkUserId,
                 role: newRole == 'Admin' ? 'org:admin' : 'org:member'
             })
         }
     }
 
-    return updatedPerson
+    return updatedMembership
 }
